Vietnam-LLM/chronos-bolt-base-demo/requirements.txt
# Core
pandas>=2.2
numpy>=1.26
matplotlib>=3.8

# Data
yfinance>=0.2.40

# Chronos (Forecasting pipeline)
chronos-forecasting>=1.3.0

# Torch - install CPU by default; users with CUDA can override
# On Windows/Linux with Python 3.13, pip will resolve the right wheel.
torch>=2.4

# Optional performance
accelerate>=0.33
